Added a check for velocity core

+++ entrypoint.sh
@@ -1,8 +1,10 @@
 #!/bin/bash
 
-cp /tmp/eula.txt /minecraft/eula.txt
-
-/utils/setup_properties.sh
 /utils/download_utils/dist/main
+
+if ! unzip -l "/minecraft/server.jar" | awk '{print $4}' | grep -qx "default-velocity.toml"; then
+    cp /tmp/eula.txt /minecraft/eula.txt
+    /utils/setup_properties.sh
+fi
 
 exec /utils/start_server.sh
